delete old MachineFluidTank impl and renames
