+++ src/machines/java/com/enderio/machines/common/recipe/TankRecipe.java
@@ -4,7 +4,7 @@ import com.enderio.EnderIO;
 import com.enderio.core.common.recipes.EnderRecipe;
 import com.enderio.machines.common.blockentity.FluidTankBlockEntity;
 import com.enderio.machines.common.init.MachineRecipes;
-import com.enderio.machines.common.io.fluid.MachineTank;
+import com.enderio.machines.common.io.fluid.MachineFluidTank;
 import com.google.gson.JsonObject;
 import net.minecraft.ResourceLocationException;
 import net.minecraft.core.RegistryAccess;
@@ -101,14 +101,14 @@ public class TankRecipe implements EnderRecipe<TankRecipe.Container> {
 
     public static class Container extends RecipeWrapper {
 
-        private final MachineTank fluidTank;
+        private final MachineFluidTank fluidTank;
 
-        public Container(IItemHandlerModifiable inv, MachineTank fluidTank) {
+        public Container(IItemHandlerModifiable inv, MachineFluidTank fluidTank) {
             super(inv);
             this.fluidTank = fluidTank;
         }
 
-        public MachineTank getFluidTank() {
+        public MachineFluidTank getFluidTank() {
             return fluidTank;
         }
     }
